requirements.txt
beautifulsoup4==4.5.1
cycler==0.10.0
Django==1.10.3
jieba==0.38
lxml==3.6.4
matplotlib==1.5.3
numpy==1.11.2
pandas==0.19.2
pika==0.10.0
Pillow==3.4.2
pymongo==3.3.1
pyparsing==2.1.10
python-dateutil==2.6.0
python-docx==0.8.6
pytrends==3.2.0
pytz==2016.7
requests==2.12.3
selenium==3.0.2
six==1.10.0
wordcloud==1.2.1
xlrd==1.0.0
